+++ core/translatedGraphicsFunction.h
@@ -584,8 +584,8 @@ public:
 	virtual void delete_shader(unsigned int&) = 0;
 
 	virtual void gen_buffer(unsigned int*) = 0;
-	virtual void bind_buffer(int, unsigned int*) = 0;
-	virtual void buffer_data(int, unsigned int, const void*, int) = 0;
+	virtual void bind_buffer(unsigned int, unsigned int*) = 0;
+	virtual void buffer_data(unsigned int, long, const void*, unsigned int) = 0;
 	virtual void delete_buffer(unsigned int*) = 0;
 
 	virtual void gen_vertex_array(unsigned int*) = 0;
